+++ src/lib.rs
@@ -2,6 +2,13 @@
 //! based on the [`embedded-hal`](https://github.com/japaric/embedded-hal) traits.
 //! 
 //! This driver allows you to:
+//! - Display single digits or characters, also at a selected position
+//! - Display text, although some characters may not be available (see display documentation)
+//! - Clear the display
+//! - Show the current I2C address
+//! - Change the I2C address (experimental function)
+//! - Display time in HH.MM format
+//! - Display temperature or humidity, with settable lower/upper threshold
 //! 
 //!## The device
 //! The TWI 7-segment Display is an easy to use 4-digit 7-segment display that is controlled using the TWI (I2C compatible) protocol.
@@ -16,28 +23,129 @@
 //! [examples]: https://github.com/nebelgrau77/akafugu_twidisplay-rs/tree/main/examples
 //! 
 //! ### Initialization
+//! A new instance of the device is created as follows:
 //! 
+//! ```rust
+//! use akafugu_twidisplay::*:
+//! 
+//! let mut akafugu = TWIDisplay::new(i2c, DEFAULT_ADDRESS);
+//! ```
+//! 
+//! The default address is 0x12. If the address was changed with the `set_address()` function,
+//! the new address must be used after a power down-power up sequence.  
+//! 
+//! 
+//! ### Main functions
+//! 
+//! Display can be cleared with the following command:
+//! ```rust
+//! akafugu.clear_display().unwrap();
+//! ```
+//! 
+//! Digits and/or characters can either be simply sent to display, or displayed at defined positions.
+//! 
+//! ```rust
+//! // display digit '7' at position 2 (positions are 0,1,2,3 from left to right)
+//! akafugu.display_digit(2, 7).unwrap();
+//! // display character 'P' at position 3
+//! akafugu.display_char(3,'P').unwrap();
+//! ```
+//! 
+//! If a digit/character is just sent to the display, it will appear according to the selected mode
+//! (scroll or rotate) - please see the documentation. 
+//! 
+//! ```rust
+//! akafugu.send_char('A').unwrap();
+//! akafugu.send_char('B').unwrap();
+//! akafugu.send_char('C').unwrap();
+//! akafugu.send_char('D').unwrap();
+//! ```
+//! 
+//! This will display `ABCD`. 
+//! 
+//! ```rust
+//! akafugu.send_char('E').unwrap();
+//! ```
+//! 
+//! Depending on the selected mode the display will show now:
+//! * in SCROLL mode: 'BCDE' 
+//! * in ROTATE mode: 'EBCD'
+//! 
+//! Text can be sent to display as string literals:
+//! 
+//! ```rust
+//! akafugu.send_text("HELLO LOOP PULL CALL").unwrap();
+//! ```
+//! 
+//! Numbers from 0-9999 range can be displayed with the following function:
+//! ```rust
+//! akafugu.display_number(1234).unwrap();
+//! ```
+//! _NOTE_: Numbers will be displayed with leading zeroes, e.g. `0023`. 
+//! 
+//! ### Control functions
+//! 
+//! Display mode can be changed as follows:
+//! 
+//! ```rust
+//! akafugu.set_mode(Mode::Scroll).unwrap(); // default mode is `Rotate`
+//! ```
+//! 
+//! Brightness can be set between 0 and 255, where 127 is approx. 50% brightness.
+//! ```rust 
+//! akafugu.set_brightness(200).unwrap();
+//! ```
+//! 
+//! The I2C address of the device can be changed from the default 0x12 as follows:
+//! ```rust
+//! akafugu.set_address(0x20).unwrap();
+//! ```
+//! 
+//! The new address will be active after a power down, power up sequence.
+//! 
+//! __NOTE:__ According to the documentation, the allowed range of addresses is 0x00-0x7F,
+//! but addresses including and over 0x40 don't seem to work correctly, even though
+//! they are correctly displayed. In such case 0x00 must be used to access the device and change the address again. 
+//! For this reason in this driver the address setting is restricted to 0x00-0x39 range.
+//! 
+//! To show the current I2C address use the following command:
+//! ```rust
+//! akafugu.display_address().unwrap();
+//! ```
+//! The same can be achieved by simply connecting only the VCC and GND pins of the display.
+//! 
+//! ### Convenience functions
+//! The driver has three additional functions, that can be useful for clock or sensor applications.
+//! 
+//! #### Display time
+//! Time is displayed in HH.MM format, with the central dot displayed or not:
+//! 
+//! ```rust
+//! 
+//! // get time from the clock 
+//! let (hours, minutes, seconds) = some_rtc_function();
+//! 
+//! // blink the dot: on if number of seconds is even, otherwise off
+//! if seconds % 2 == 0 {
+//!     akafugu.display_time(hours, minutes, true).unwrap()
+//! } else {
+//!     akafugu.display_time(hours, minutes, false).unwrap()
+//! }
+//! ```
+//! 
+//! #### Display temperature
+//! 
+
 
 // TO DO:
 /*
 
-- make the driver initialization with a specific address from an enum
 
-Functions to add:
-
-GENERAL:
-- sending digits
-- sending strings
 
 COMMANDS:
-DONE: - brightness setting, takes one argument 0-255 DONE
-- address setting: takes one argument, actually can be anything between 0-127, so it can't be an enum, must be a number, with a match 
-DONE: - clearing the display
-- scroll/rotate mode: takes one argument, 0 or 1 (rotate/scroll) - this one can use an enum
 - dots setting
 - getting firmware revision
 - getting number of digits
-DONE: - display current I2C address
 
 not tested by me yet: 
 - custom character
@@ -45,24 +153,22 @@ not tested by me yet:
 
 - time displaying - doesn't seem to be working
 
-FUNCTIONS:
-DONE - send digit
-- send character
-- send string
-- display time (hh:mm:ss with a blinking dot)
-DONE - display temperature (with a choice of C or F unit)
 
 
 TO DO:
 
-ADD READ FUNCTION 
-ADD READING FIRMWARE / VERSION
-ADD SETTING I2C ADDRESS
-DONE: ADD UNIVERSAL DATA DISPLAY FUNCTION
-MODIFY TEMPERATURE FUNCTION
-ADD HUMIDITY FUNCTION
-ADD STRING PRINTING
+FUNCTIONS:
+
+- READ() FUNCTION 
+- READING FIRMWARE / VERSION
+- SETTING I2C ADDRESS
+- SETTING DOTS WITH BOOLEANS
+
 USE '-' FOR DIGITS IF INCORRECT (DOES IT MAKE SENSE?)
+
+OTHER:
+- CUSTOM CHARACTERS
+
 
 */
 
@@ -74,7 +180,7 @@ USE '-' FOR DIGITS IF INCORRECT (DOES IT MAKE SENSE?)
 
 use embedded_hal as hal;
 
-use hal::blocking::i2c::{Write, WriteRead, Read};
+use hal::blocking::i2c::{Write, WriteRead};
 
 /// All possible errors in this crate
 #[derive(Debug)]
@@ -198,24 +304,25 @@ where
         Ok(())
     }
 
-    /*
+    
 
     // NEED TO TEST MORE: TRIED WITH VALUE 0x69, CORRECTLY DISPLAYED A105 ON POWER-UP
-    // BUT DID NOT RESPOND ON 0x69 I2C ADDRESS
-    // RESPONDED ONLY WITH ADDRESS 0
+    // BUT DID NOT RESPOND ON 0x69 I2C ADDRESS, RESPONDED ONLY WITH ADDRESS 0
+    // SEEMS TO WORK OK UP TO 0x39
+    // -- USE ADDRESS 0x00 TO RESET IN CASE OF PROBLEMS
 
     /// Set I2C address, defaults to 0x12
     pub fn set_address(&mut self, address: u8) -> Result<(), Error<E>> {
         //let mut dev_address = DEFAULT_ADDRESS;
         match address {
-            a if a < 0x7f => self.write(&[Register::I2C_ADDRESS_SETTING, a])?,
+            //a if a < 0x7f => self.write(&[Register::I2C_ADDRESS_SETTING, a])?,
+            a if a < 0x40 => self.write(&[Register::I2C_ADDRESS_SETTING, a])?,
             _ => (),
             }
         Ok(())
         }
     
-     */
-
+    
 
     /// Show the current I2C address on the display
     pub fn display_address(&mut self) -> Result<(), Error<E>> {
@@ -342,7 +449,7 @@ where
     }
 
     /// Display data with units (temperature, humidity) and defined thresholds
-    pub fn display_data(&mut self, 
+    fn display_data(&mut self, 
                     data: i16, unit: char, 
                     lo_thresh: Option<i16>, hi_thresh: Option<i16>, 
                     min_val: i16, max_val: i16) -> Result<(), Error<E>> {